requirements.txt
flask
flask-cors
flask-bcrypt
textblob
deep-translator
langdetect
google-generativeai
